requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.0.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
deprecation==2.1.0
dill==0.3.8
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
postgrest==0.15.0
pydantic==2.6.1
pydantic_core==2.16.2
pylint==3.0.3
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.6
supafunc==0.3.3
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.2.0
websockets==11.0.3
